requirements.txt
adrf==0.1.9
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-property==0.2.2
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kombu==5.4.2
openai==1.57.4
packaging==24.2
prompt_toolkit==3.0.50
psycopg==3.2.3
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
